Add the ability to force-load on new versions
This change will allow for a developer to force the plugin to load anyways should the Minecraft version be unsupported by Foundation (untested). This will allow more advanced users of the library to update their servers without any unnecessary dependence on the Mineacademy team. The primary use case of this feature would be to allow servers and networks to update their codebases as fast as possible should new snapshots be released.