Externalize comments engine, finalize new settings draft
After a week of work, it's finally almost done. Remaining:
1. Excessive testing with all premium plugins.
2. Comments and javadocs.

Still subject to change, do not use yet. If you are a student of MineAcademy you will receive a free training video once it is ready. If not, good news, we plan on releasing videos to YouTube later this year!